chore: refactored the code for better maintainability
